requirements.txt
absl-py==0.7.0
alembic==1.0.8
aniso8601==6.0.0
astor==0.7.1
attrs==19.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
Django==2.2.1
django-cors-headers==2.5.3
django-filter==1.1.0
django-json-field==0.5.7
djangorestframework==3.9.3
djangorestframework-filters==0.11.1
forex-python==1.3
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.16.2
pandas==0.24.1
protobuf==3.6.1
psycopg2==2.7.7
pymongo==3.7.2
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
serpy==0.3.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.0
sqlparse==0.3.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
